[Feat] #16 - 반응형 로직 디테일 추가
